+++ src/main/java/org/frontear/mixins/impl/MixinEntityRenderer.java
@@ -1,4 +1,4 @@
-package org.frontear.mixins;
+package org.frontear.mixins.impl;
 
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiIngame;
